Update EasyTransferI2C.cpp
Added a function to be used in the slave-sender's request handler. I didn't modify your great cheksum system just crammed the message in a buffer and sent all in one write, because in the request handler only one write is allowed. Works great but it is limited to the buffer size this way. (Although that can be increased) Thanks for the great library.